requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
asttokens==2.0.8
backcall==0.2.0
certifi @ file:///C:/b/abs_4f5wo627a3/croots/recipe/certifi_1663615677642/work/certifi
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
dj-database-url==1.0.0
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-cors-headers==3.13.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.0.0
gunicorn==20.1.0
imageio==2.22.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
pywin32==304
pyzmq==24.0.0
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
stack-data==0.5.0
tifffile==2022.8.12
tornado==6.2
traitlets==5.4.0
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
wcwidth==0.2.5
whitenoise==6.2.0
wincertstore==0.2
wheel==0.37.1
Django==4.1.0